Move to e2e-testing folder
