+++ packages/server-wallet/e2e-testing/test-data/roles.json
